requirements.txt
Flask==2.3.2
pymongo[srv]==4.3.3
python-dotenv==1.0.0
gunicorn==20.1.0
